+++ battlenet-api-browser/main.js
@@ -1,26 +1,82 @@
-const { app, BrowserWindow } = require('electron/main')
+const { app, BrowserWindow, protocol } = require('electron/main')
+const fs = require('node:fs');
+const path = require('node:path');
+const url = require('url');
 
+let win;
+
+const rootFile = 'dist/battlenet-api-browser/browser/index.html';
+console.log("file: "+rootFile);
+const absRoot = path.resolve(rootFile);
+console.log("absFile: "+absRoot);
+
+/**
+ * createWindow function to create our main window
+ */
 const createWindow = () => {
-  const win = new BrowserWindow({
+  win = new BrowserWindow({
     width: 800,
     height: 600
   })
 
-  win.loadFile('dist/battlenet-api-browser/browser/index.html')
+  /**
+   * Catch load failures (e.g. Electron's refresh, which doesn't include index.html on the end)  
+   * and redirect them to the index
+   */
+  win.webContents.on('did-fail-load', () => { 
+    win.loadURL(absRoot); 
+  });
+
+
+  /**
+   * Catch redirect events from oauth
+   */
+  win.webContents.on('will-redirect', (event)=>{
+    if (event.url.startsWith("http:")) {
+      //console.log("intercepting redirect "+event.url);
+      event.preventDefault();  
+      const thisUrl = new URL(event.url);
+      const params = new URLSearchParams(thisUrl.search);   
+      const code = params.get('code');
+      const state = params.get('state');
+      win.loadURL(absRoot+"?code="+code+"&state="+state); 
+    }
+  });
+
+  /**
+   * Begin by loading our index file
+   */
+  win.loadFile(absRoot);
 }
 
-app.whenReady().then(() => {
-  createWindow()
-
-  app.on('activate', () => {
-    if (BrowserWindow.getAllWindows().length === 0) {
-      createWindow()
-    }
-  })
-})
-
+/**
+ * Standard shutdown procedure for electron - terminate app when all windows are closed.
+ * Otherwise the Electron process will just hang around window-less
+ */
 app.on('window-all-closed', () => {
+  //except on OSX
   if (process.platform !== 'darwin') {
     app.quit()
+  } 
+});
+
+/**
+ * When application is activated, open the main window if we don't have one
+ */
+app.on('activate', () => {
+  if (BrowserWindow.getAllWindows().length === 0) {
+    createWindow()
   }
 })
+
+
+
+/**
+ * When application is ready, open the main window and set up our http interceptor
+ */
+app.whenReady().then(() => {
+  createWindow();
+})
+
+
+