+++ battlenet-api-browser/src/browse/list-detail-host/list-detail-host.component.ts
@@ -3,7 +3,7 @@ import { apiClientService } from '../../services/apiclient.service';
 import { Subscription } from 'rxjs';
 import { ScrollingModule } from '@angular/cdk/scrolling';
 import { UserdataService } from '../../services/userdata.service';
-import { IApiDataDoc } from '../../model/datastructs';
+import { IApiDataDoc, topDataStruct } from '../../model/datastructs';
 import { ActivatedRoute } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { MatButtonModule } from '@angular/material/button';
@@ -113,19 +113,25 @@ export class ListDetailHostComponent implements OnInit, OnDestroy {
   {
     if (this.mode == this.Mode.Master)
     {
-      this.masterList!.reload(this.apiClient).then((list)=>{
-        this.userData.dataRefreshedEmitter.emit();
-      });
+      if (this.masterList?.canReload())
+      {
+        this.masterList?.reload(this.apiClient).then((list)=>{
+          this.userData.dataRefreshedEmitter.emit();
+        });
+      }
     }
     else
     {
-      this.masterList!.reloadItem(this.apiClient, this.id).then((rec)=>{
-        this.detailItem = rec;
-        this.userData.setCurrent(this.masterList!, this.detailItem!);
-        this.mode = ListDetailHostComponentMode.Detail;
-        this.detailInputs = { 'data': this.detailItem! };               
-        this.userData.dataRefreshedEmitter.emit();
-      });
+      if (this.masterList?.canReloadItems())
+      {
+        this.masterList?.reloadItem(this.apiClient, this.id).then((rec)=>{
+          this.detailItem = rec;
+          this.userData.setCurrent(this.masterList!, this.detailItem!);
+          this.mode = ListDetailHostComponentMode.Detail;
+          this.detailInputs = { 'data': this.detailItem! };               
+          this.userData.dataRefreshedEmitter.emit();
+        });
+      }
     }
   }
 
@@ -146,9 +152,11 @@ export class ListDetailHostComponent implements OnInit, OnDestroy {
     this.id = undefined;
     var idstr = this.route.snapshot.paramMap.get('id');
 
-    //find reference from string passed in route data
+    //find reference from string array passed in route data
     this.data = this.route.snapshot.data as ListDetailHostComponentData;
-    this.masterList = this.getValueByKey(this.data.list, this.userData.data.apiData);
+    const section = this.getValueByKey(this.data.list[0], this.userData.data.apiData) as topDataStruct;
+    this.masterList = section.getData(this.data.list[1]);
+    //this.masterList = this.getValueByKey(this.data.list, this.userData.data.apiData);
     
     if (this.masterList!.stringKey)
     {