+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 ## Changelog
 
 #### Current-Dev (awaiting release number)
+(todo)
 
-#### Upgrade Notes
+#### v0.3d
+
+##### Upgrade Notes
 - Storage of API key has changed, if upgrading you'll need to re-input your Client ID and Secret.
 
-#### New Features
+##### New Features
 - Added Character Profiles data
 - Added Guild Profiles data
 - Added a simple query cache
+- The command "bna" can now be used in the browser console for debugging purposes
 - Extension support (partial)
 - "Rebuild Index" item available on tools tab where applicable
 - Settings allows changing connection type if additional types are available (requires extensions)
@@ -21,6 +25,7 @@
 - Records now delete properly when using the tools "clear this data type" button.
 - API 404s no longer hang jobs in the job queue.
 - Rendering of tables within tables in "(Data)" view has been reduced to reduce complexity hanging browsers
+- OAuth now works properly on Electron
 
 #### v0.2d